requirements.txt
streamlit
google.cloud
googletrans
os